+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/case-flag-summary-list/case-flag-summary-list.component.spec.ts
@@ -15,6 +15,8 @@ describe('CaseFlagSummaryListComponent', () => {
   let mockRpxTranslationService: any;
   const updateFlagHeaderText = 'Update flag for';
   const addFlagHeaderText = 'Add flag to';
+  const updateSupportHeaderText = 'Update support for';
+  const addSupportHeaderText = 'Add support to';
 
   const flagDetailDisplay = {
     partyName: 'Rose Bank',
@@ -25,7 +27,8 @@ describe('CaseFlagSummaryListComponent', () => {
       path: [{ id: '', value: 'Reasonable adjustment' }],
       hearingRelevant: false,
       flagCode: 'FL1',
-      status: 'Active'
+      status: 'Active',
+      flagStatusReasonChange: 'Change of status'
     } as FlagDetail
   } as FlagDetailDisplay;
 
@@ -158,9 +161,61 @@ describe('CaseFlagSummaryListComponent', () => {
     expect(summaryListValues[3].textContent).toContain(flag.flagDetail.status);
   });
 
+  it('should display the flag summary for a flag with comments, as part of the Create Case Flag journey for an external user', () => {
+    component.flagForSummaryDisplay = flagDetailDisplay;
+    component.displayContextParameter = CaseFlagDisplayContextParameter.CREATE_EXTERNAL;
+    fixture.detectChanges();
+    const addUpdateFlagHeaderTextElement = nativeElement.querySelector('dt');
+    expect(addUpdateFlagHeaderTextElement.textContent).toContain(addSupportHeaderText);
+    const summaryListValues = nativeElement.querySelectorAll('dd.govuk-summary-list__value');
+    expect(summaryListValues[0].textContent).toContain(flagDetailDisplay.partyName);
+    expect(summaryListValues[1].textContent).toContain(flagDetailDisplay.flagDetail.name);
+    expect(summaryListValues[2].textContent).toContain(flagDetailDisplay.flagDetail.flagComment);
+    expect(summaryListValues[3].textContent).toContain(flagDetailDisplay.flagDetail.status);
+  });
+
+  it('should display the flag summary for a flag with comments, as part of the Create Case Flag journey with v2.1 enabled', () => {
+    component.flagForSummaryDisplay = flagDetailDisplay;
+    component.displayContextParameter = CaseFlagDisplayContextParameter.CREATE_2_POINT_1;
+    fixture.detectChanges();
+    const addUpdateFlagHeaderTextElement = nativeElement.querySelector('dt');
+    expect(addUpdateFlagHeaderTextElement.textContent).toContain(addFlagHeaderText);
+    const summaryListValues = nativeElement.querySelectorAll('dd.govuk-summary-list__value');
+    expect(summaryListValues[0].textContent).toContain(flagDetailDisplay.partyName);
+    expect(summaryListValues[1].textContent).toContain(flagDetailDisplay.flagDetail.name);
+    expect(summaryListValues[2].textContent).toContain(flagDetailDisplay.flagDetail.flagComment);
+    expect(summaryListValues[3].textContent).toContain(flagDetailDisplay.flagDetail.status);
+  });
+
   it('should display the flag summary for a flag with comments, as part of the Manage Case Flags journey', () => {
     component.flagForSummaryDisplay = flagDetailDisplay;
     component.displayContextParameter = CaseFlagDisplayContextParameter.UPDATE;
+    fixture.detectChanges();
+    const addUpdateFlagHeaderTextElement = nativeElement.querySelector('dt');
+    expect(addUpdateFlagHeaderTextElement.textContent).toContain(updateFlagHeaderText);
+    const summaryListValues = nativeElement.querySelectorAll('dd.govuk-summary-list__value');
+    expect(summaryListValues[0].textContent).toContain(flagDetailDisplay.partyName);
+    expect(summaryListValues[1].textContent).toContain(flagDetailDisplay.flagDetail.name);
+    expect(summaryListValues[2].textContent).toContain(flagDetailDisplay.flagDetail.flagComment);
+    expect(summaryListValues[3].textContent).toContain(flagDetailDisplay.flagDetail.status);
+  });
+
+  it('should display the flag update comments in the summary, as part of the Manage Case Flags journey for an external user', () => {
+    component.flagForSummaryDisplay = flagDetailDisplay;
+    component.displayContextParameter = CaseFlagDisplayContextParameter.UPDATE_EXTERNAL;
+    fixture.detectChanges();
+    const addUpdateFlagHeaderTextElement = nativeElement.querySelector('dt');
+    expect(addUpdateFlagHeaderTextElement.textContent).toContain(updateSupportHeaderText);
+    const summaryListValues = nativeElement.querySelectorAll('dd.govuk-summary-list__value');
+    expect(summaryListValues[0].textContent).toContain(flagDetailDisplay.partyName);
+    expect(summaryListValues[1].textContent).toContain(flagDetailDisplay.flagDetail.name);
+    expect(summaryListValues[2].textContent).toContain(flagDetailDisplay.flagDetail['flagStatusReasonChange']);
+    expect(summaryListValues[3].textContent).toContain(flagDetailDisplay.flagDetail.status);
+  });
+
+  it('should display the flag comments in the summary, as part of the Manage Case Flags journey with v2.1 enabled', () => {
+    component.flagForSummaryDisplay = flagDetailDisplay;
+    component.displayContextParameter = CaseFlagDisplayContextParameter.UPDATE_2_POINT_1;
     fixture.detectChanges();
     const addUpdateFlagHeaderTextElement = nativeElement.querySelector('dt');
     expect(addUpdateFlagHeaderTextElement.textContent).toContain(updateFlagHeaderText);